+++ cmd/server/config.go
@@ -62,9 +62,10 @@ func readOrGenerateConfig(path string) (configJSON string, writable bool, err er
 func generateConfigFile(path string) (configJSON string, err error) {
 	// The default site configuration.
 	defaultSiteConfig := schema.SiteConfiguration{
-		SecretKey:        string(mustCryptoRand()),
-		AuthAllowSignup:  false,
-		AuthProvider:     "builtin",
+		SecretKey: string(mustCryptoRand()),
+		AuthProviders: []schema.AuthProviders{
+			{Builtin: &schema.BuiltinAuthProvider{Type: "builtin"}},
+		},
 		UpdateChannel:    "release",
 		MaxReposToSearch: 50,