+++ schema/forwardcompat_auth_providers.go
@@ -1,9 +1,0 @@
-package schema
-
-// AuthProviders
-type AuthProviders struct {
-	Builtin       *BuiltinAuthProvider
-	Saml          *SAMLAuthProvider
-	Openidconnect *OpenIDConnectAuthProvider
-	HttpHeader    *HTTPHeaderAuthProvider
-}